Refactor: Update ROS integration, local assets, and UI for development
This commit introduces several changes to facilitate development and testing, including using a local ROS URL, disabling certain ROS subscriptions for mock data, adding local font assets, and making minor UI adjustments.

**Key Changes:**

*   **ROS Configuration & ViewModels:**
    *   **`lib/config/ros_constants.dart`:** Changed `rosUrl` to `ws://192.168.133.56:9090` (dynalog ethernet) for local development.
    *   **`lib/view_models/boot_check_view_model.dart`:**
        *   Commented out `_topic!.subscribe(_responseHandler)`.
        *   Now initializes `state` with a default `BootCheckResponse` indicating "OK" status.
    *   **`lib/view_models/operation_view_model.dart`:**
        *   Commented out `_topicCurrentMode!.subscribe(_handler)`.
        *   Now initializes `state` to `'delivery'`.
    *   **`lib/view_models/table_view_model.dart`:**
        *   Constructor now initializes the state with a default list of tables (1 to 10).
    *   **`lib/presentation/screens/delivery_screens/delivery_main_screen.dart`:**
        *   Commented out calls to `requestTableList` and `getTableList` in `initState`.

*   **Asset Management:**
    *   **`pubspec.yaml`:**
        *   Removed `device_preview` dependency.
        *   Added `assets/google_fonts/` to assets.
    *   **`lib/main.dart`:**
        *   Removed `device_preview` and `window_manager` related code.
        *   Set `GoogleFonts.config.allowRuntimeFetching = false`.
    *   Added new font files to `assets/google_fonts/`:
        *   `Inter-Bold.ttf`, `Inter-Medium.ttf`, `Inter-SemiBold.ttf`
        *   `Nunito-Black.ttf`, `Nunito-Bold.ttf`, `Nunito-SemiBold.ttf`
        *   `Roboto-Bold.ttf`, `Roboto-Medium.ttf`, `Roboto-SemiBold.ttf`

*   **Model Update:**
    *   **`lib/models/battery_status.dart`:**
        *   Changed `percentage` field type from `int?` to `double?`.
        *   Updated `batteryPercentage` getter to return `percentage!.toInt()`.

*   **UI Adjustment:**
    *   **`lib/presentation/screens/delivery_screens/delivery_main_screen.dart`:**
        *   In the "No tables found" message, changed `fontWeight` of the text to `FontWeight.w400`.

+++ lib/models/battery_status.dart
@@ -8,7 +8,7 @@ BatteryStatus batteryStatusFromJson(String str) =>
 String batteryStatusToJson(BatteryStatus data) => json.encode(data.toJson());
 
 class BatteryStatus {
-  int? percentage;
+  double? percentage;
   int? state;
   double? voltage;
   double? current;
@@ -38,7 +38,7 @@ class BatteryStatus {
     if(percentage == null) {
       return 0;
     }
-    return percentage!;
+    return percentage!.toInt();
   }
 
   BatteryStatusType get batteryStatusType {
